Recalculate: the DC block's seven per-iteration ids, 84.03 -> 95.27
The DC LOCAL-SCOPE SWEEP's first real payout.  THillFortWindow::Recalculate's
Dreamcast procedure block declares SEVEN id variables at hillfortwindow.cpp
:271..277, one statement each, computed unconditionally before the
`s.type != CREATURE_NONE && s.count > 0` gate at :279 - six of them named in
the CodeView local list, the seventh (the portrait id) register-allocated into
r4 and consumed at :284.  The SH4 constants identify them exactly: :271 K1,
:272 K1+7, :273 K1+14 with K1 = CREATURE_PORTRAIT_1_ID = 205, then :274
GOLD_COST_1_ID, :275 +7 = RES_ICON_1_ID, :276 RES_COST_1_ID, :277 +21 =
UPGRADE_BUTTON_1_ID.

Writing all seven and DROPPING the carried `id` is worth 84.0331 -> 94.0227 on
its own.  An earlier note had "reintroducing all six beside the current
i/reference scores 79.60" - that probe kept `id`, which is the whole
difference; VC6 CSEs the seven back onto retail's single [ebp-8] home and
re-derives `+0xe` from it, so the widget-id half of the recorded induction
anchor bias disappears.

The affordability verdict is the DC's CanAfford CALL, not the hand-written
loop with its `goto have_state`: :332/:337/:345 are
`if (!IsBaseCreature) state=0; else if (CanAfford(s.cost, resources)) state=1;
else state=2;`.  That is worth a further 94.0227 -> 95.2727 and it puts
retail's two arms back in retail's layout order.

CanAfford takes `const long*` in the DC roster, so `TUpgradeSlot::cost` is
`long[7]` and the `get_upgrade_cost` out-parameter family widened with it
(recruit.h/.cpp plus the four caller-local `cost` arrays in viewarmywindow,
game and philai) - byte-flat tree-wide, only the mangled name moves
(PAH -> PAJ) and the ratchet row renamed with it.

Residual 95.2727: three size-only blocks behind the one surviving anchor bias
(retail's slot cursor is &slot[0].count, ours &slot[0].type).

+++ src/philai.cpp
@@ -935,7 +935,7 @@ __forceinline long value_of_hill_fort(const hero* current_hero,
 {
     int funds[NUM_RESOURCES];
     memcpy(funds, gpCurrentPlayer->resources, sizeof(funds));
-    int cost[NUM_RESOURCES];
+    long cost[NUM_RESOURCES];
     long value = 0;
 
     for (int slot = 0; slot < armyGroup::ARMY_GROUP_SLOT_COUNT; ++slot) {
@@ -3258,7 +3258,7 @@ int ValueOfStables(const hero* current_hero, long* move_cost)
 VA(0x00527bb0, 0x11D)  // anchor-callee, dc 0x111630
 void AI_visit_hill_fort(hero* current_hero)
 {
-    int cost[NUM_RESOURCES];
+    long cost[NUM_RESOURCES];
 
     for (int i = 0; i < armyGroup::ARMY_GROUP_SLOT_COUNT; i++) {
         TCreatureType creature = current_hero->army.armyTypes[i];
